+++ src/ej_conversations/tests/conftest.py
@@ -2,6 +2,59 @@ import json
 import pytest
 
 from django.test.client import Client
+from django.contrib.auth.models import AnonymousUser
+
+from ej_conversations import create_conversation
+from ej_users.models import User
+
+
+@pytest.fixture
+def user(db):
+    user = User.objects.create_user('email@server.com', 'password')
+    user.board_name = 'testboard'
+    user.save()
+    return user
+
+
+@pytest.fixture
+def conversation(db, user):
+    return create_conversation(text='test', title='title', author=user, is_promoted=True)
+
+
+@pytest.fixture
+def comment(db, conversation, user):
+    return conversation.create_comment(user, 'content', 'approved')
+
+
+@pytest.fixture
+def vote(db, user, comment):
+    return comment.vote(author=user, choice='agree')
+
+
+@pytest.fixture
+def post_request(rf):
+    request = rf.post('')
+    request.user = AnonymousUser()
+    return request
+
+
+@pytest.fixture
+def request_(rf):
+    request = rf.get('')
+    request.user = AnonymousUser()
+    return request
+
+
+@pytest.fixture
+def request_with_user(rf, user):
+    request = rf.get('/testboard/')
+    request.user = user
+    return request
+
+
+@pytest.fixture
+def mk_user(db, email='default@user.com', is_staff=False):
+    return User.objects.create_user(email, '1234', is_staff=is_staff)
 
 
 @pytest.fixture
