+++ src/ej_conversations/tests/test_views.py
@@ -1,52 +1,11 @@
-import pytest
-from django.contrib.auth.models import AnonymousUser
 from pytest import raises
 from django.http import HttpResponseServerError, Http404
+from django.contrib.auth.models import AnonymousUser
 
-from ej_conversations import create_conversation
 from ej_conversations.models import Comment, FavoriteConversation
 from ej_conversations.models.utils import votes_counter
 from ej_conversations.routes import conversations, comments, admin
 from ej_users.models import User
-
-
-@pytest.fixture
-def post_request(rf):
-    request = rf.post('')
-    request.user = AnonymousUser()
-    return request
-
-
-@pytest.fixture
-def request_(rf):
-    request = rf.get('')
-    request.user = AnonymousUser()
-    return request
-
-
-@pytest.fixture
-def request_with_user(rf, user):
-    request = rf.get('/testboard/')
-    request.user = user
-    return request
-
-
-@pytest.fixture
-def user(db):
-    user = User.objects.create_user('email@server.com', 'password')
-    user.board_name = 'testboard'
-    user.save()
-    return user
-
-
-@pytest.fixture
-def conversation(db, user):
-    return create_conversation(text='test', title='title', author=user, is_promoted=True)
-
-
-@pytest.fixture
-def comment(db, conversation, user):
-    return conversation.create_comment(user, 'content', 'approved')
 
 
 class TestConversationBase:
